Add click-through for WOSB
- skip the auto delete 'Draft' application for now

+++ src/main/java/gov/sba/automation/AutomationUtils.java
@@ -22,7 +22,15 @@ public class AutomationUtils {
         wdriver = browserUtils.getWaitDriver();
         driver = browserUtils.getWebDriver();
     }
+	
+    public WebDriverWait getWdriver() {
+		return wdriver;
+	}
 
+	public WebDriver getDriver() {
+		return driver;
+	}
+    
     public void login(String username, String password) {
         driver.findElement(By.cssSelector(".button-full")).click();
         wdriver.until(ExpectedConditions.presenceOfElementLocated(By.cssSelector("#user_email")));
@@ -49,10 +57,14 @@ public class AutomationUtils {
         // Now let's click the program link
         CoreUtils.locateAndClick(driver, locator);
         
-        logger.debug("FYI: about to call deleteDrafProgram()");
-        // TODO: delete the 'Draft' application if any?
-        CoreUtils.deleteDraftProgram(driver, programName);
-        Thread.sleep(5000);
+        boolean deleteDraft = false;
+        
+        if (deleteDraft) {
+        	logger.debug("FYI: about to call deleteDrafProgram()");
+        	// TODO: delete the 'Draft' application if any?
+        	CoreUtils.deleteDraftProgram(driver, programName);
+        	Thread.sleep(5000);
+        }
         
         // Get the program name
         String programDesc = CoreUtils.lookupProgram(programName);
@@ -68,10 +80,44 @@ public class AutomationUtils {
         // Then we click Continue button
         CoreUtils.clickContinue(driver);
     }
+    
+    public void clickThroughWosb(WebDriver driver) {
+    	CoreUtils.yesOrNo(driver, "no", new int[] {198});
+    	CoreUtils.clickContinue(driver);
 
+    	CoreUtils.yesOrNo(driver, "no", new int[] {199});
+    	CoreUtils.clickContinue(driver);
+
+    	CoreUtils.yesOrNo(driver, "no", new int[] {201});
+    	CoreUtils.clickContinue(driver);
+
+    	CoreUtils.yesOrNo(driver, "no", new int[] {211});
+    	CoreUtils.comment(driver, 211);
+
+    	CoreUtils.yesOrNo(driver, "no", new int[] {212});
+    	CoreUtils.comment(driver, 212);
+    	CoreUtils.clickContinue(driver);
+
+    	CoreUtils.yesOrNo(driver, "no", new int[] {213});
+    	CoreUtils.clickContinue(driver);
+    	
+    	CoreUtils.yesOrNo(driver, "no", new int[] {214, 215, 216});
+    	CoreUtils.comment(driver, 216);
+    	CoreUtils.clickContinue(driver);
+
+    	CoreUtils.yesOrNo(driver, "no", new int[] {217, 218, 219, 220, 221, 222});
+    	CoreUtils.comment(driver, 222);
+    	CoreUtils.clickContinue(driver);
+    	
+    	CoreUtils.yesOrNo(driver, "no", new int[] {223});
+    	CoreUtils.clickContinue(driver);
+    }
+    
     public static void main(String... args) throws Exception {
         AutomationUtils app = new AutomationUtils();
         app.startNewProgram("wosb");
+        
+        app.clickThroughWosb(app.getDriver());
         System.out.println("Done!");
     }
 }
